+++ core/src/main/scala/org/apache/spark/status/api/v1/api.scala
@@ -391,11 +391,3 @@ case class ThreadStackTrace(
     val blockedByLock: String,
     val holdingLocks: Seq[String])
 
-class ExecutionData (val id : Long,
-    val status: String,
-    val description: String,
-    val submissionTime: String,
-    val duration: String,
-    val runningJobs: Seq[Int],
-    val successJobs: Seq[Int],
-    val failedJobs: Seq[Int])
